fix: update octokit/throttling plugin (#4)
* chore: upgrade octokit dependencies

The main goal here is to upgrade `@octokit/plugin-throttling` to `5`, so that the secondary rate limit wait time is increased, wich hopefully fixes https://github.com/intuit/auto/issues/2458. This required other dependencies to be upgraded. The `resolutions` for `@octokit/plugin-rest-endpoint-methods` is necessary, because without it it pulls in `10.0.0` of `@octokit/types` which isn't compatible with the other libraries.

* fix: another attempt to avoid using 10.x types

* fix: typo

* fix: git dependencies

Finally found a combination of @octokit versions that all use @octokit/types 9.x

* chore: downgrade for node compatibility reasons

* fix: downgrade for node compatibility

* fix: downgrade for node compatibility

---------

Co-authored-by: Patrick Ruhkopf <patrick@ruhkopf.me>